+++ CHANGELOG.md
@@ -2,6 +2,40 @@
 
 ## [Unreleased]
 
+## [5.14.0](https://github.com/near/near-sdk-rs/compare/near-sdk-v5.13.0...near-sdk-v5.14.0) - 2025-05-14
+
+### Other
+
+- updates near-workspaces to 0.20 version ([#1358](https://github.com/near/near-sdk-rs/pull/1358))
+- updates near-* dependencies to 0.30 release ([#1356](https://github.com/near/near-sdk-rs/pull/1356))
+
+## [5.13.0](https://github.com/near/near-sdk-rs/compare/near-sdk-v5.12.0...near-sdk-v5.13.0) - 2025-05-05
+
+### Added
+
+- Added BLS12-381 curve operations support in near-sys and exposed in near_sdk::env ([#1346](https://github.com/near/near-sdk-rs/pull/1346))
+
+### Fixed
+
+- BLS12-381 pairing check return value comparison ([#1352](https://github.com/near/near-sdk-rs/pull/1352))
+- Fixed the tokenization of of ContractMetadata, where it can lead to invalid meta if `link` and `version` is not provided ([#1349](https://github.com/near/near-sdk-rs/pull/1349))
+
+### Other
+
+- update `near-workspaces`, `cargo-near-build`in examples and core tests  to avoid `wasm-opt` compile ([#1350](https://github.com/near/near-sdk-rs/pull/1350))
+
+## [5.12.0](https://github.com/near/near-sdk-rs/compare/near-sdk-v5.11.0...near-sdk-v5.12.0) - 2025-04-10
+
+### Added
+
+- nep330 result wasm path field  ([#1344](https://github.com/near/near-sdk-rs/pull/1344))
+
+### Other
+
+- Fixed linting warnings (clippy Rust 1.86) ([#1345](https://github.com/near/near-sdk-rs/pull/1345))
+- added example how to use `promise_yield_create` and `promise_yield_resume` ([#1133](https://github.com/near/near-sdk-rs/pull/1133))
+- updated near-workspaces in examples ([#1337](https://github.com/near/near-sdk-rs/pull/1337))
+
 ## [5.11.0](https://github.com/near/near-sdk-rs/compare/near-sdk-v5.10.0...near-sdk-v5.11.0) - 2025-03-22
 
 ### Added